# Better/faster rust code