It improves Rust performance from ~3100ms on my machine to ~2450ms

asm: 2400ms
c: 2700ms
c++: 2750ms